Move StrictMode out of options package
