+++ src/main/java/com/maddyhome/idea/vim/newapi/IjVimEditor.kt
@@ -54,7 +54,7 @@ import com.maddyhome.idea.vim.helper.updateCaretsVisualPosition
 import com.maddyhome.idea.vim.helper.vimChangeActionSwitchMode
 import com.maddyhome.idea.vim.helper.vimKeepingVisualOperatorAction
 import com.maddyhome.idea.vim.helper.vimLastSelectionType
-import com.maddyhome.idea.vim.options.helpers.StrictMode
+import com.maddyhome.idea.vim.helper.StrictMode
 import org.jetbrains.annotations.ApiStatus
 
 @ApiStatus.Internal
