+++ vim-engine/src/main/kotlin/com/maddyhome/idea/vim/api/VimCaret.kt
@@ -21,7 +21,7 @@ import com.maddyhome.idea.vim.helper.exitVisualMode
 import com.maddyhome.idea.vim.helper.inBlockSubMode
 import com.maddyhome.idea.vim.helper.inSelectMode
 import com.maddyhome.idea.vim.helper.inVisualMode
-import com.maddyhome.idea.vim.options.helpers.StrictMode
+import com.maddyhome.idea.vim.helper.StrictMode
 import com.maddyhome.idea.vim.register.Register
 import javax.swing.KeyStroke
