+++ vim-engine/src/main/kotlin/com/maddyhome/idea/vim/helper/StrictMode.kt
@@ -6,7 +6,7 @@
  * https://opensource.org/licenses/MIT.
  */
 
-package com.maddyhome.idea.vim.options.helpers
+package com.maddyhome.idea.vim.helper
 
 import com.maddyhome.idea.vim.api.globalOptions
 import com.maddyhome.idea.vim.api.injector
