+++ vim-engine/src/main/kotlin/com/maddyhome/idea/vim/option/StrictMode.kt
@@ -12,7 +12,7 @@ import com.maddyhome.idea.vim.api.globalOptions
 import com.maddyhome.idea.vim.api.injector
 import com.maddyhome.idea.vim.options.OptionConstants
 
-@Deprecated("Please use StrictMode from com.maddyhome.idea.vim.options.helpers", replaceWith = ReplaceWith("com.maddyhome.idea.vim.options.helpers.StrictMode"))
+@Deprecated("Please use StrictMode from com.maddyhome.idea.vim.options.helpers", replaceWith = ReplaceWith("com.maddyhome.idea.vim.helper.StrictMode"))
 object StrictMode {
   @JvmName("assertTrue")
   fun assert(condition: Boolean, message: String) {
